Create Change The Text
To get text on screen right click in the little box that says create and choose UI then text.
After that zero the transform so that the text appears in the middle of the screen.
There is a text box. Write what you want in it and then when you press play that is what will appear.
Add this script if you want to change the text.
It is not very useful in it's present form but with a few if statments added it could be.